requirements.txt
alembic==0.8.9
amqp==2.1.3
beautifulsoup4==4.4.1
billiard==3.5.0.2
blinker==1.4
celery==4.0.1
cffi==1.7.0
chardet==2.3.0
click==6.6
coverage==4.2
cryptography==1.4
dominate==2.3.1
enum34==1.1.6
Flask-Bootstrap==3.3.7.0
Flask-DebugToolbar==0.10.0
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.0.2
Flask-Moment==0.5.1
flask-mongoengine==0.8.2
Flask-Redis==0.3.0
Flask-Script==2.0.5
Flask-SocketIO==2.8.1
Flask-SQLAlchemy==2.1
Flask-WTF==0.13.1
gevent==1.1.2
gevent-websocket==0.9.5
greenlet==0.4.11
idna==2.1
ipaddress==1.0.16
itsdangerous==0.24
jieba==0.38
Jinja2==2.8
kombu==4.0.1
lxml==3.5.0
Mako==1.0.6
MarkupSafe==0.23
mongoengine==0.11.0
ndg-httpsclient==0.4.2
nose==1.3.7
Pillow==3.4.2
psycopg2==2.6.1
pyasn1==0.1.9
pycparser==2.14
pymongo==3.4.0
pyOpenSSL==16.0.0
pyping==0.0.5
pypinyin==0.16.0
python-docx==0.8.6
python-editor==1.0.3
python-engineio==1.1.0
python-socketio==1.6.1
pytz==2016.10
qiniu==7.1.0
qrcode==5.3
redis==2.10.5
requests==2.12.3
six==1.10.0
SQLAlchemy==1.1.4
ua-parser==0.7.2
user-agents==1.0.1
vine==1.1.3
visitor==0.1.3
Werkzeug==0.11.11
WTForms==2.1
xlrd==1.0.0
XlsxWriter==0.9.4
